requirements.txt
langchain
wikipedia
langchain-ollama
crewai
crewai[tools]
pydantic
python-dotenv
langgraph